AI_backend/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
docker==7.0.0
idna==3.6
packaging==24.0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1

djangorestframework==3.14.0
jsonfield==3.1.0
psycopg2-binary==2.9.3
gunicorn==20.1.0